Move has_triton to top level triton utils so that dynamo can also access it without creating cyclic dependencies
[ghstack-poisoned]
