+++ test/inductor/test_mmdecomp.py
@@ -5,7 +5,6 @@ import unittest
 from typing import List, Tuple, Union
 
 import torch
-from torch._inductor.utils import has_triton
 from torch.testing._internal.common_device_type import instantiate_device_type_tests
 from torch.testing._internal.common_nn import NNTestCase
 from torch.testing._internal.common_utils import (
@@ -14,6 +13,7 @@ from torch.testing._internal.common_utils import (
     run_tests,
     TEST_CUDA,
 )
+from torch.utils._triton import has_triton
 
 
 default_atol = {
